Address detached folder icon review comments
